mv func related to files in loadouts.go and file.go and do the same for login funcs and login.go
